Update to 1.4.1 release of origin
Includes results of "gofmt -w". Modify the code to work against
origin 1.4.1. The major change here is that scaling must be
done against the deploymentconfig and not the replication
controller.

+++ clusters/containers/containers.go
@@ -84,7 +84,7 @@ type OContainerPort struct {
 func ContainerPort(name string, port int) *OContainerPort {
 	cp := OContainerPort{}
 	cp.Name = name
-	cp.ContainerPort.ContainerPort = port
+	cp.ContainerPort.ContainerPort = int32(port)
 	cp.ContainerPort.Protocol = kapi.ProtocolTCP
 	return &cp
 }
@@ -100,7 +100,7 @@ func (cp *OContainerPort) SetName(name string) *OContainerPort {
 }
 
 func (cp *OContainerPort) HostPort(port int) *OContainerPort {
-	cp.ContainerPort.HostPort = port
+	cp.ContainerPort.HostPort = int32(port)
 	return cp
 }
